change everything to double from float

+++ hand_traits.h
@@ -15,22 +15,22 @@ enum class size : int { small, big };
 template<size H>
 struct hand_traits
 {
-    static constexpr float pinch_min = 1.0;
-    static constexpr float pinch_max = 1.0;
+    static constexpr double pinch_min = 1.0;
+    static constexpr double pinch_max = 1.0;
 };
 
 template<>
 struct hand_traits<size::small>
 {
-    static constexpr float pinch_min = 1.0;
-    static constexpr float pinch_max = 1.0;
+    static constexpr double pinch_min = 1.0;
+    static constexpr double pinch_max = 1.0;
 };
 
 template<>
 struct hand_traits<size::big>
 {
-    static constexpr float pinch_min = 1.0;
-    static constexpr float pinch_max = 1.0;
+    static constexpr double pinch_min = 1.0;
+    static constexpr double pinch_max = 1.0;
 };
 
 }
